// src/app/core/api/model/userResponseDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserResponseDto { 
    id?: number;
    username?: string;
    name?: string;
    roleId?: number;
    roleName?: string;
    branchId?: number;
    tenantId?: number;
    isActive?: boolean;
}